Fix CVE for rubyzip
https://nvd.nist.gov/vuln/detail/CVE-2018-1000544

rubyzip gem rubyzip version 1.2.1 and earlier contains a
Directory Traversal vulnerability in Zip::File component that can
result in write arbitrary files to the filesystem. This attack appear
to be exploitable via If a site allows uploading of .zip files , an
attacker can upload a malicious file that contains symlinks or files
with absolute pathnames "../" to write arbitrary files to the
filesystem.