Use Rails URL helper, not direct links to S3
Direct links to S3 expire every week, but Rails URLs don't.

In development and test we don't use S3, so I added a workaround
to use the local service with the host set to localhost:3000.

This will technically work and make the tests pass.
Realistically we will never install an app from localhost.