Remove duplicated code between specs
The goal of cancel_deploy_spec.rb is to test that you can click
on "Cancel" for a scheduled build and change it from "scheduled"
to "canceled".

We can reuse most of the code from the upload_spec.rb file.

Additionally, we don't need to be picky with the Timeslot that we
create. We can use `first_or_create!` because tests simply need one.