Ensure package file exists to extract manifest
In Rails 6, the internals of Active Storage have changed: files
are only uploaded after a model is saved. Specifically the are
uploaded inside an `after_commit_update`, see:

https://github.com/rails/rails/pull/33303

Since our code also uses `after_commit_update` to run the job
that analyzes that file to extract the manifest, we have to
make sure that our own callback is run **after** the Active Storage
one. Adding `prepend: true` accomplishes that.