Ensure jobs are run in specific tests
In Rails 5.x, having this setting in config/environments/test.rb

```
config.active_job.queue_adapter     = :inline
```

was enough to tell Active Job to **run** the jobs during test.

In Rails 6, the behavior has changed because of
https://github.com/rails/rails/pull/33849
that uses the Test Adapter by default.

Therefore we need to specify *which* tests to run with InlineAdapter.