Extract display of form errors into helper
This commit has two goals.

In first place, it removes code repetition and logic from the view.
Rather than having, for instance `form.file_field` twice in the view,
it now only happens once, wrapped by a helper that correctly adds the
red highlighting and messages in case of error.

In second place, it adds flexibility in letting us display errors
related to *multiple* attributes next to a single field. This will
be valuable in the context of the following commit.